# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Quantum FX Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/earnings-chart:
    get:
      operationId: getEarningsChart
      tags: [dashboard]
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7D, 30D, 90D]
      responses:
        "200":
          description: Earnings chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChartPoint"
...
  /cashier/deposit:
    post:
      operationId: createDeposit
      tags: [cashier]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInput"
      responses:
        "201":
          description: Deposit initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cashier/withdraw:
    post:
      operationId: createWithdrawal
      tags: [cashier]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawInput"
      responses:
        "201":
          description: Withdrawal initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cashier/transactions:
    get:
      operationId: listTransactions
      tags: [cashier]
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [all, deposit, withdrawal]
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
...
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # Trade
...
    ChartPoint:
      type: object
      properties:
        date:
          type: string
        value:
          type: number
      required: [date, value]
...
    DepositInput:
      type: object
      properties:
        amount:
          type: number
        paymentMethod:
          type: string
        walletAddress:
          type: ["string", "null"]
      required: [amount, paymentMethod]
    WithdrawInput:
      type: object
      properties:
        amount:
          type: number
        paymentMethod:
          type: string
        walletAddress:
          type: string
      required: [amount, paymentMethod, walletAddress]
    Transaction:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        amount:
          type: number
        status:
          type: string
        paymentMethod:
          type: string
        createdAt:
          type: string
        walletAddress:
          type: ["string", "null"]
      required: [id, type, amount, status, paymentMethod, createdAt]
...